bot/pyproject.toml
[project]
name = "lms-bot"
version = "0.1.0"
description = "LMS Telegram Bot"
requires-python = "==3.14.2"
dependencies = [
    "aiogram>=3.20",
    "httpx==0.28.1",
    "pydantic-settings==2.12.0",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=22.7",
]
